split code-action DTOs into dedicated microcrate
