+++ Cargo.toml
@@ -42,6 +42,7 @@ members = [
     "crates/perl-lsp-launcher",
     "crates/perl-lsp-navigation",
     "crates/perl-lsp-completion",
+    "crates/perl-lsp-code-action-types",
     "crates/perl-lsp-code-actions",
     "crates/perl-lsp-cancellation",
     "crates/perl-lsp-limits",
@@ -135,6 +136,7 @@ allow = [
     "perl-semantic-analyzer",
     "perl-lsp-diagnostics",
     "perl-lsp-rename",
+    "perl-lsp-code-action-types",
     "perl-lsp-code-actions",
     "perl-lsp-completion",
     "perl-lsp-inlay-hints",
@@ -290,6 +292,7 @@ perl-lsp-launcher = { path = "crates/perl-lsp-launcher", version = "0.10.0" }
 perl-lsp-navigation = { path = "crates/perl-lsp-navigation", version = "0.10.0" }
 perl-lsp-rename = { path = "crates/perl-lsp-rename", version = "0.10.0" }
 perl-lsp-completion = { path = "crates/perl-lsp-completion", version = "0.10.0" }
+perl-lsp-code-action-types = { path = "crates/perl-lsp-code-action-types", version = "0.10.0" }
 perl-lsp-code-actions = { path = "crates/perl-lsp-code-actions", version = "0.10.0" }
 perl-diagnostics-codes = { path = "crates/perl-diagnostics-codes", version = "0.10.0" }
