+++ crates/perl-lsp-code-action-types/README.md
@@ -1,0 +1,3 @@
+# perl-lsp-code-action-types
+
+Shared data transfer objects used by Perl LSP code action providers.
