+++ crates/perl-lsp-code-actions/src/code_actions.rs
@@ -67,9 +67,7 @@
 
 use crate::quick_fixes;
 use crate::refactors;
-use crate::types::QuickFixDiagnostic;
-
-pub use crate::types::{CodeAction, CodeActionKind};
+use perl_lsp_code_action_types::{CodeAction, QuickFixDiagnostic};
 
 use perl_lsp_diagnostics::Diagnostic;
 use perl_parser_core::Node;
