+++ crates/perl-lsp-code-actions/src/enhanced/extract_subroutine.rs
@@ -1,6 +1,6 @@
 //! Extract subroutine code action
 
-use crate::types::{CodeAction, CodeActionEdit, CodeActionKind};
+use perl_lsp_code_action_types::{CodeAction, CodeActionEdit, CodeActionKind};
 use perl_lsp_rename::TextEdit;
 use perl_parser_core::ast::{Node, NodeKind, SourceLocation};
 use std::collections::HashSet;
