+++ crates/perl-lsp-code-actions/src/enhanced/loop_conversion.rs
@@ -1,6 +1,6 @@
 //! Loop conversion code actions
 
-use crate::types::{CodeAction, CodeActionEdit, CodeActionKind};
+use perl_lsp_code_action_types::{CodeAction, CodeActionEdit, CodeActionKind};
 use perl_lsp_rename::TextEdit;
 use perl_parser_core::ast::{Node, NodeKind};
