+++ crates/perl-lsp-code-actions/src/quick_fixes.rs
@@ -3,7 +3,7 @@
 //! Provides automated fixes for common Perl issues driven by diagnostic codes.
 
 use crate::ast_utils::{find_declaration_position, get_indent_at};
-use crate::types::{CodeAction, CodeActionEdit, CodeActionKind, QuickFixDiagnostic};
+use perl_lsp_code_action_types::{CodeAction, CodeActionEdit, CodeActionKind, QuickFixDiagnostic};
 use perl_lsp_rename::TextEdit;
 use perl_parser_core::SourceLocation;
